GCD example working with schema class
